backend/AI-hybrid-chatbot/requirements.txt
streamlit
pandas
sentence-transformers
faiss-cpu
requests
fastapi
uvicorn
openpyxl